moved new-card.js into components folder
